refactor: move type to shared/types folder
